requirements.txt
crewai==0.28.8
crewai[tools]==0.28.8
openai==1.12.0
beautifulsoup4==4.12.3
requests==2.31.0
gspread==6.0.0
google-auth==2.27.0
python-dotenv==1.0.1
tenacity==8.2.3
